Day11 Lab 2
I created a comprehensive website with a lot of data and different tags, but I have done what is expected in this assignment. I did not fully replicate the website I made on the first day as it was way over the requirements. If I am still expected to do so, I can add in the extra stuff.